Added separate lightweight metadata
